+++ actionpack/lib/action_view/helpers/tags/select.rb
@@ -13,7 +13,8 @@ module ActionView
         def render
           option_tags_options = {
             :selected => @options.fetch(:selected) { value(@object) },
-            :disabled => @options[:disabled]
+            :disabled => @options[:disabled],
+            :priority => @options[:priority]
           }
 
           option_tags = if grouped_choices?